backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
apscheduler==3.10.4
python-dotenv==1.0.1
gunicorn==22.0.0
sqlalchemy==2.0.31
